% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wn_col.R
\name{check_wn_col}
\alias{check_wn_col}
\title{Check the Currently Set Wavenumber Column}
\usage{
check_wn_col(show_val = FALSE)
}
\arguments{
\item{show_val}{Logical. If `TRUE`, the function returns the name of the current wavenumber column. Default is `FALSE`.}
}
\value{
If `show_val = TRUE`, returns the name of the wavenumber column (character). Otherwise, it only prints a message.
}
\description{
This function checks which column has been set as the default wavenumber column.
If no column has been set, it prints a message prompting the user to define one.
}
\examples{
set_spec_wn("Wavenumber")  # Set the wavenumber column
check_wn_col()             # Check which column is set
check_wn_col(show_val = TRUE)  # Check and return the column name

}
