% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spec_wn.R
\name{set_spec_wn}
\alias{set_spec_wn}
\title{Set the Default Wavenumber Column}
\usage{
set_spec_wn(col_name)
}
\arguments{
\item{col_name}{A string specifying the name of the wavenumber column.}
}
\value{
Invisibly returns the assigned column name.
}
\description{
Defines a default wavenumber column name to be used by other functions in the package when `wn_col` is not explicitly provided.
}
\examples{
set_spec_wn("Wavenumber")
}
\seealso{
[spec_select()]
}
