% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_bl_irls.R
\name{spec_bl_irls}
\alias{spec_bl_irls}
\title{Perform Baseline Estimation Using IRLS Method}
\usage{
spec_bl_irls(
  .data,
  wn_col = NULL,
  Wn_min = NULL,
  Wn_max = NULL,
  lambda1,
  lambda2,
  maxit = 200,
  wi = 0.05,
  is_abs = TRUE
)
}
\arguments{
\item{.data}{A data frame or tibble containing the spectral data.}

\item{wn_col}{A character string specifying the column name that contains the wavelength or wavenumber values. Default is `"Wn"`.}

\item{Wn_min}{Numeric value indicating the minimum wavenumber to filter the data.}

\item{Wn_max}{Numeric value indicating the maximum wavenumber to filter the data.}

\item{lambda1}{A numeric value for the 2nd derivative constraint for primary smoothing in the IRLS method.}

\item{lambda2}{A numeric value for the 2nd derivative constraint for secondary smoothing in the IRLS method.}

\item{maxit}{An integer specifying the maximum number of iterations for the IRLS method. Default is 200.}

\item{wi}{A numeric value representing the initial weight for the IRLS algorithm. Default is 0.05.}

\item{is_abs}{A logical value indicating whether the input data is in absorbance (`TRUE`) or transmittance (`FALSE`). Default is `TRUE`.}
}
\value{
A tibble with the estimated baseline of the spectral data, including the wavenumber column.
}
\description{
This function estimates the baseline of spectral data using the Iteratively Reweighted Least Squares (IRLS) method.
The baseline correction can be applied to absorbance or transmittance spectra depending on the `is_abs` parameter.
}
\references{
Baseline estimation performed using the `baseline` package for R.
More information can be found at: \url{https://CRAN.R-project.org/package=baseline}
}
\seealso{
[baseline::baseline()], [spec_trans2abs()]
}
