% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_blc_rubberband.R
\name{spec_blc_rubberband}
\alias{spec_blc_rubberband}
\title{Apply Rubberband Baseline Correction to Spectral Data}
\usage{
spec_blc_rubberband(
  .data,
  wn_col = NULL,
  wn_min = NULL,
  wn_max = NULL,
  segment_length = 50,
  smooth_baseline = TRUE,
  is_abs = TRUE
)
}
\arguments{
\item{.data}{A `data.frame` or `tibble` containing spectral data.}

\item{wn_col}{A character string specifying the column name for the wavelength data. Default is `"Wn"`.}

\item{wn_min}{A numeric value specifying the minimum wavelength to consider for the baseline correction.}

\item{wn_max}{A numeric value specifying the maximum wavelength to consider for the baseline correction.}

\item{segment_length}{A numeric value specifying the length of segments for finding local minima. Default is 50.}

\item{smooth_baseline}{A logical value indicating whether to smooth the baseline using spline interpolation. Default is TRUE.}

\item{is_abs}{A logical value indicating whether the data is already in absorbance. If `TRUE`, absorbance is used directly; if `FALSE`, the data is converted to absorbance before applying the baseline correction.}
}
\value{
A `tibble` with the baseline-corrected spectral data, containing the wavelength column and the corrected numeric columns.
}
\description{
This function applies a rubberband baseline correction to spectral data within a specified wavelength range.
The rubberband method fits a baseline by connecting local minima points in the spectrum.
It allows for correction of either absorbance or transmittance data.
}
\references{
Rubberband baseline correction connects local minima to estimate baseline.
}
