% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_norm_01.R
\name{spec_norm_01}
\alias{spec_norm_01}
\title{Normalize Spectra Data to [0,1] Range}
\usage{
spec_norm_01(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{data.frame, the input data set}

\item{wn_col}{character, the name of the column representing the Wavenumber, default is "Wn"}
}
\value{
a data.frame with normalized values of the predictors in [0,1] range, Wn column is preserved.
}
\description{
Normalize Spectra Data to [0,1] Range
}
