% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_norm_var.R
\name{spec_norm_var}
\alias{spec_norm_var}
\title{Normalize the Spectra Data by Feature Scaling}
\usage{
spec_norm_var(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{A data.frame that contains the Spectra Data.}

\item{wn_col}{The name of the column that contains the wavenumber of the spectra data.}
}
\value{
A normalized Spectra Data.
}
\description{
Normalize the Spectra Data by Feature Scaling so that all the predictors have mean = 0 and variance = 1.
}
