% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_smooth_sga.R
\name{spec_smooth_sga}
\alias{spec_smooth_sga}
\title{Smooth Spectra Using the S-Golay Filter}
\usage{
spec_smooth_sga(.data, wn_col = "Wn", window = 15, forder = 4, degree = 0)
}
\arguments{
\item{.data}{A data frame with spectra data, where the first column is
the wavenumber (Wn) column.}

\item{wn_col}{The name of the wavenumber column in the data frame.
Default is "Wn".}

\item{window}{The size of the sliding window for the S-Golay filter.
The size of the window must be an odd number.
Default is 15.}

\item{forder}{The order of the polynomial to fit in the S-Golay filter.
The order must be smaller than the window size.
Default is 4.}

\item{degree}{The degree of the polynomial to fit in the S-Golay filter.
The degree must be less than or equal to the forder.
Default is 0.}
}
\value{
A data frame with the smoothed spectra, where the first column is
  the wavenumber column and the rest of the columns are the smoothed
  spectra values.
}
\description{
This function smooths spectra using the Savitzky-Golay (S-Golay) filter.
It applies local polynomial regression to the data, fitting a polynomial over
a sliding window of `window` points to smooth the spectra.
}
\examples{
# Load example spectra data
data(spectra_data)

# Smooth spectra using the S-Golay filter
smooth_spectra <- spec_smooth_sga(spectra_data)

# Plot the smoothed spectra
plot(smooth_spectra)

}
