% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_ball.R
\name{smooth_baseline}
\alias{smooth_baseline}
\title{Baseline Smoothing}
\usage{
smooth_baseline(baseline, ws)
}
\arguments{
\item{baseline}{Vector with the baseline}

\item{ws}{Smoothing window width}
}
\value{
Smoothed vector
}
\description{
Auxiliary function to smooth the baseline using moving average
}
