% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_abs2trans.R
\name{spec_abs2trans}
\alias{spec_abs2trans}
\title{Convert Absorbance Spectra to Transmittance Spectra}
\usage{
spec_abs2trans(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{A data.frame or tibble containing the absorbance spectra.}

\item{wn_col}{The name of the column in the data containing the wavenumber. Default is "Wn".}
}
\value{
A tibble containing the wavenumber and corresponding transmittance spectra.
}
\description{
Converts a spectra in absorbance to transmittance by applying the formula: T = 10^(2-A).
}
\examples{
# Load example data
library(tidyspec)
data(spectra_abs)

# Convert to transmittance
spectra_trans <- spec_abs2trans(spectra_abs)

}
