% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_blc_rollingBall.R
\name{spec_blc_rollingBall}
\alias{spec_blc_rollingBall}
\title{Remove baseline drift from a spectral data set}
\usage{
spec_blc_rollingBall(
  .data,
  wn_col = "Wn",
  Wn_min,
  Wn_max,
  wm,
  ws,
  is_abs = TRUE
)
}
\arguments{
\item{.data}{A data frame with columns Wn and one or more spectral data}

\item{Wn_min}{Minimum Wn value to consider}

\item{Wn_max}{Maximum Wn value to consider}

\item{wm}{Window size for moving average calculation}

\item{ws}{Window size for baseline determination}

\item{is_abs}{A logical indicating if the input data is in absorbance units}
}
\value{
A data frame with Wn and the corrected spectral data
}
\description{
Remove baseline drift from a spectral data set
}
\examples{
spec_blc_rollingBall(data, 2000, 4000, wm=15, ws=30)

}
