% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_diff.R
\name{spec_diff}
\alias{spec_diff}
\title{Title: Spectral Derivative}
\usage{
spec_diff(.data, wn_col = "Wn", degree = 1)
}
\arguments{
\item{.data}{A tibble or data.frame containing the spectral data.}

\item{wn_col}{Character string representing the column name for the Wn data.}

\item{degree}{Numeric value for the derivative degree, default is 1.}
}
\value{
A tibble or data.frame with columns for each derivative of the input data and the Wn column.
}
\description{
The `spec_diff` function calculates the derivative of a spectral data set. The default derivative degree is 1, but this can be changed to any positive integer. The resulting data set will contain columns for each derivative and the Wn column.
}
