% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_norm_01.R
\name{spec_norm_01}
\alias{spec_norm_01}
\title{Normalize Spectral Data to the [0, 1] Range}
\usage{
spec_norm_01(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{A `data.frame` or `tibble` containing spectral data.}

\item{wn_col}{A character string specifying the column name for the wavelength data. Default is `"Wn"`.}
}
\value{
A `tibble` with the normalized spectral data, containing the wavelength column and the normalized numeric columns.
}
\description{
This function normalizes the numeric spectral data in each column to the [0, 1] range, preserving the wavelength column.
}
