% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_norm_minmax.R
\name{spec_norm_minmax}
\alias{spec_norm_minmax}
\title{Normalize Data using Min-Max Scaler}
\usage{
spec_norm_minmax(.data, wn_col = "Wn", min = 0, max = 1)
}
\arguments{
\item{.data}{data.frame. The input data.}

\item{wn_col}{character. The column representing the dependent variable (Wavenumber). Default is "Wn".}

\item{min}{numeric. The minimum value to be used in scaling. Default is 0.}

\item{max}{numeric. The maximum value to be used in scaling. Default is 1.}
}
\value{
A data.frame with only the dependent variable column and the normalized predictors.
}
\description{
Normalize the numeric predictors using Min-Max Scaler, with custom values of minimum and maximum scaling.
}
\examples{
spec_norm_minmax(mtcars, "mpg", 0, 10)
}
