% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_norm_var.R
\name{spec_norm_var}
\alias{spec_norm_var}
\title{Standardize Spectral Data to Unit Variance}
\usage{
spec_norm_var(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{A `data.frame` or `tibble` containing spectral data.}

\item{wn_col}{A character string specifying the column name for the wavelength data. Default is `"Wn"`.}
}
\value{
A `tibble` with the standardized spectral data, containing the wavelength column and the standardized numeric columns.
}
\description{
This function standardizes the numeric spectral data in each column to have a mean of 0 and a standard deviation of 1 (unit variance), while preserving the wavelength column.
}
