% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_select.R
\name{spec_select}
\alias{spec_select}
\title{Select Specific Columns in a Spectral Data Frame}
\usage{
spec_select(.data, wn_col = NULL, ...)
}
\arguments{
\item{.data}{A data frame containing spectral data.}

\item{wn_col}{A string specifying the column name that represents the wavenumber values. If NULL, the function will attempt to retrieve a default value set with `set_spec_wn()`.}

\item{...}{Additional column names to be selected, provided as unquoted variable names.}
}
\value{
A data frame containing only the selected columns.
}
\description{
This function selects user-specified columns from a spectral dataset, always ensuring that the wavenumber column (`wn_col`) is included.
}
\seealso{
[dplyr::select()], [set_spec_wn()]
}
