% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_smartplot.R
\name{spec_smartplot}
\alias{spec_smartplot}
\title{Plot Spectral Data}
\usage{
spec_smartplot(
  .data,
  wn_col = "Wn",
  type = c("absorbance", "transmittance"),
  xdir = c("reverse", "standard"),
  geom = c("point", "line"),
  xmin = 400,
  xmax = 4000,
  alpha = 0.8
)
}
\arguments{
\item{.data}{data.frame to plot}

\item{wn_col}{character string, the column name of the wavenumber}

\item{type}{character string, plot type "absorbance" or "transmittance"}

\item{xdir}{character string, plot direction "reverse" or "standard"}

\item{geom}{character string, plot style "point" or "line"}

\item{xmin}{numeric, lower bound of the x-axis}

\item{xmax}{numeric, upper bound of the x-axis}

\item{alpha}{numeric, plot transparency between 0 and 1}
}
\value{
A ggplot object with the specified parameters
}
\description{
Plot spectral data with the given parameters using ggplot2. The spectral data is plotted as either absorbance or transmittance, using either a line or a point plot, with a selectable color map, and can be reversed on the x-axis.
}
\examples{
spec_smartplot(.data = iris, wn_col = "Sepal.Length",
               type = "transmittance", xdir = "standard",
               geom = "line", xmin = 3, xmax = 8, alpha = .8)
}
