% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_smartplotly.R
\name{spec_smartplotly}
\alias{spec_smartplotly}
\title{Plot Spectral Data}
\usage{
spec_smartplotly(
  .data,
  wn_col = "Wn",
  type = c("absorbance", "transmittance"),
  xdir = c("reverse", "standard"),
  geom = c("point", "line"),
  xmin = 400,
  xmax = 4000,
  alpha = 0.8
)
}
\arguments{
\item{.data}{A data frame containing the spectra data.}

\item{wn_col}{A character string specifying the name of the wavenumber column in the data.}

\item{type}{A character string that specifies the type of spectra (absorbance or transmittance).}

\item{xdir}{A character string that specifies the direction of the x-axis (reverse or standard).}

\item{geom}{A character string that specifies the type of plot geometry to use (point or line).}

\item{xmin}{Numeric value that specifies the minimum x-axis value.}

\item{xmax}{Numeric value that specifies the maximum x-axis value.}

\item{alpha}{A numeric value specifying the alpha value for the plot.}
}
\value{
A interactive plotly plot of the spectra data.
}
\description{
This function is used to create a interactive plot of spectral data. The user can specify the type of spectra (absorbance or transmittance), the wavenumber column, and the minimum and maximum x-axis values.
}
\examples{
spec_smartplotly(spectra_data,wn_col="Wn",type="absorbance",xmin=400,xmax=4000)
}
