% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_trans2abs.R
\name{spec_trans2abs}
\alias{spec_trans2abs}
\title{Convert transmittance spectra to absorbance spectra}
\usage{
spec_trans2abs(.data, wn_col = "Wn")
}
\arguments{
\item{.data}{A data frame with the spectra data}

\item{wn_col}{A string specifying the name of the wavenumber column}
}
\value{
A data frame with the same wavenumber column and absorbance values in
place of the transmittance values. Infinite values are filtered out.
}
\description{
Given a data frame with transmittance spectra, this function converts it to
absorbance spectra. The result is a new data frame with the same wavenumber
column and absorbance values in place of the transmittance values.
}
\examples{

spec_trans2abs(.data = spectra_data, wn_col = "Wn")

}
